requirements.txt
openai>=1.45.0
tavily-python>=0.5.0
tenacity>=8.5.0
pydantic>=2.9.0
loguru>=0.7.2
python-dotenv>=1.0.1
